% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_vis.R
\name{add_gender}
\alias{add_gender}
\title{Adds a variable called \code{gender} to a data.frame}
\usage{
add_gender(df, gender_var)
}
\arguments{
\item{df}{the data.frame}

\item{gender_var}{the optional name of the gender variable}
}
\value{
a data.frame with the variable \code{gender}
}
\description{
Adds a variable called \code{gender} to a data.frame
}
