% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_total}
\alias{add_total}
\title{Add row with total}
\usage{
add_total(data)
}
\arguments{
\item{data}{data.frame or tibble}
}
\value{
df with additional row with all numeric
columns summed up.
}
\description{
Brutally sums up all numeric columns and appends
to data.
}
