% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{age}
\alias{age}
\title{Time between dates}
\usage{
age(from_date, to_date, unit = "years")
}
\arguments{
\item{from_date}{start date of interval}

\item{to_date}{end date for interval}

\item{unit}{Unit to use, passed to lubridate::time_length
(days, weeks, months, years)}
}
\description{
Calculates time between two dates.
}
