% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_vis.R
\name{aggregate_vis}
\alias{aggregate_vis}
\title{Aggregate a VIS indicator}
\usage{
aggregate_vis(
  df,
  measure_id,
  indicator_function,
  date_var,
  unit_var,
  county_var,
  register_id,
  gender_var = NULL,
  ...
)
}
\arguments{
\item{df}{a data.frame}

\item{measure_id}{the measure ID}

\item{indicator_function}{the function responsible for defining and
aggregating the indicator}

\item{date_var}{the name of the date variable}

\item{unit_var}{the name of the unit variable}

\item{county_var}{the name of the county variable}

\item{register_id}{the register ID}

\item{gender_var}{the name of the gender variable, if \code{NULL} the variable
\code{SubjectKey} must be present in \code{df}}

\item{...}{passed to \code{indicator_function}}
}
\value{
an aggregated data.frame with VIS appropriate names and metadata
}
\description{
Creates an aggregated indicator ready for VIS with appropiate variable names
and metadata
}
\details{
The config defining the indicator is loaded automatically from the
Kvalitetsindikatorkatalog API ensuring the indicators are always up-to-date.

Note that only the reported periods and organizational levels as well as
historic indicator validity is dynamically applied. The actual definition
of the indicator can change in the config and should be then changed in
\code{indicator_function} to reflect this.

It's important to note that all relevant filtering on, for example, gender,
and care type has to be performed before passing \code{df} to this function.

The \code{indicator_function} has to accept the following arguments \code{group_cols},
\code{obfuscate_data}, \code{aggregate}, \code{marginal_cols} and is almost surely defined
in the \link[rcdm:rcdm]{rcdm::rcdm} package.
}
