% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andel_missing.R
\name{andel_missing}
\alias{andel_missing}
\title{Derives birthdate from Swedish social security number}
\usage{
andel_missing(data, vars = names(data), ...)
}
\arguments{
\item{x}{Swedish social security number}
}
\description{
Derives birthdate from Swedish social security number
}
