% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_aggregate.R, R/api_metadata.R
\name{api_form_count}
\alias{api_form_count}
\alias{api_get}
\alias{api_RegisterID2RegisterName}
\alias{api_FormID2FormName}
\alias{api_FormID2RegisterID}
\alias{api_register}
\alias{api_register_meta}
\alias{api_register_forms}
\alias{api_register_units}
\alias{api_form}
\alias{api_form_meta}
\alias{api_form_questions}
\alias{api_map}
\title{Get API data for register, forms and questions}
\usage{
api_form_count(FormID)

api_get(x)

api_RegisterID2RegisterName(RegisterID)

api_FormID2FormName(FormID)

api_FormID2RegisterID(FormID)

api_register(RegisterID)

api_register_meta(RegisterID)

api_register_forms(RegisterID)

api_register_units(RegisterID)

api_form(FormID)

api_form_meta(FormID)

api_form_questions(FormID, all = FALSE, only_mapped = FALSE)

api_map(var, FormID)
}
\arguments{
\item{x}{character. Relevant part to paste into URL (mostly for internal use)}

\item{RegisterID, FormID}{Register and form ID numbers.
See https://medicor.github.io/metaviewer/app.html}

\item{all}{logical. Should all columns be included?}

\item{only_mapped}{logical. Should only questions with mapped domains be included?}

\item{var}{character. Column name in table.}
}
\description{
Access open data from the Stratum API.
}
\section{Functions}{
\itemize{
\item \code{api_form_count()}: Count number of registrations by form

\item \code{api_RegisterID2RegisterName()}: Convert RegisterID to RegisterName

\item \code{api_FormID2FormName()}: Convert FormID to FormName

\item \code{api_FormID2RegisterID()}: Get RegisterID for FormID

\item \code{api_register()}: All data for register from API

\item \code{api_register_meta()}: Meta data for register

\item \code{api_register_forms()}: All forms in register

\item \code{api_register_units()}: All units in register

\item \code{api_form()}: All data for register from API

\item \code{api_form_meta()}: Meta data for form

\item \code{api_form_questions()}: All questions in form

\item \code{api_map()}: get map table attribute for variable in form

}}
\references{
https://stratum.docs.apiary.io/
}
