% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{birthdate}
\alias{birthdate}
\title{Derives birthdate from Swedish social security number}
\usage{
birthdate(x)
}
\arguments{
\item{x}{Swedish social security number}
}
\description{
Derives birthdate from Swedish social security number
}
