% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{check_censor_rate}
\alias{check_censor_rate}
\title{Checks that data is not censored to an usually high degree}
\usage{
check_censor_rate(df, overall_rate = 0.25, measure_rate = 0.75)
}
\arguments{
\item{df}{VIS data.frame}

\item{overall_rate}{the overall censor rate to warn at}

\item{measure_rate}{the per-MeasureID censor rate to warn at}
}
\value{
vector of error messages
}
\description{
\itemize{
\item Ensures that the overall censor rate does not exceed \code{overall_rate}
\item Ensures that each MeasureID is not censored to a degree above
\code{measure_rate}
}
}
