% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{check_censored}
\alias{check_censored}
\title{Checks that all censored rows have values removed}
\usage{
check_censored(df)
}
\arguments{
\item{df}{VIS data.frame}
}
\value{
vector of error messages
}
\description{
\itemize{
\item Ensures that Value, Rate, ConfidenceIntervalLower and
ConfidenceIntervalHigher are missing if a ReasonCode is present
}
}
