% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{check_ci}
\alias{check_ci}
\title{Checks that all confidence intervals are correct}
\usage{
check_ci(df)
}
\arguments{
\item{df}{VIS data.frame}
}
\value{
vector of error messages
}
\description{
\itemize{
\item Ensures that no ConfidenceIntervalLower is larger than
ConfidenceIntervalHigher
\item Ensures that ConfidenceIntervalLower and ConfidenceIntervalHigher always
appear together
}
}
