% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{check_counties_and_units}
\alias{check_counties_and_units}
\title{Checks that all unit and county variables are correct}
\usage{
check_counties_and_units(df)
}
\arguments{
\item{df}{VIS data.frame}
}
\value{
vector of error messages
}
\description{
\itemize{
\item Ensures that RegionOrganisationId, RegionName and at least one of
UnitHSAID, HospitalHSAID, and HospitalOrganisationId as well as UnitName and HospitalName is not
missing
\item Ensures that each unit/hospital HSA/PAR-id belongs to at most one
unit/hospital name
\item Ensures that each unit/hospital name has exactly one HSA/PAR-id
\item Ensures that each combination of unit/hospital name and HSA/PAR-id belongs to
exactly one county
}
}
