% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{check_countryname}
\alias{check_countryname}
\title{Checks that CountryName only contains the value "Riket"}
\usage{
check_countryname(df)
}
\arguments{
\item{df}{VIS data.frame}
}
\value{
vector of error messages
}
\description{
Checks that CountryName only contains the value "Riket"
}
