% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{check_dates}
\alias{check_dates}
\title{Checks if all dates are correct}
\usage{
check_dates(df)
}
\arguments{
\item{df}{VIS data.frame}
}
\value{
vector of error messages
}
\description{
\itemize{
\item Ensures that no PeriodReportedStartDate is greater than
PeriodReportedEndDate
\item Ensures that all PeriodReportedEndDate have passed
\item Ensures that no PeriodReportedStartDate or PeriodReportedEndDate are
missing
}
}
