% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{check_duplicate_rows}
\alias{check_duplicate_rows}
\title{Checks that no rows are duplicated}
\usage{
check_duplicate_rows(df)
}
\arguments{
\item{df}{VIS data.frame}
}
\value{
vector of error messages
}
\description{
\itemize{
\item Ensures that each combination of PeriodReportedStartDate,
PeriodReportedEndDate, MeasureID, unit/hospital name, and unit/hospital
HSA/PAR-id appear exactly once
}
}
