% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{check_instruction}
\alias{check_instruction}
\title{Checks if the first cell in the row contains a KEEP or DELETE existing
instruction}
\usage{
check_instruction(df)
}
\arguments{
\item{df}{VIS data.frame}
}
\value{
vector of error messages
}
\description{
Checks if the first cell in the row contains a KEEP or DELETE existing
instruction
}
