% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{check_measureid}
\alias{check_measureid}
\title{Checks that MeasureIDs have a value}
\usage{
check_measureid(df)
}
\arguments{
\item{df}{VIS data.frame}
}
\value{
vector of error messages
}
\description{
Checks that MeasureIDs have a value
}
