% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{check_na}
\alias{check_na}
\title{Utility function to ensure that \code{var} contains no missing values}
\usage{
check_na(df, var, var_name = var)
}
\arguments{
\item{df}{VIS data.frame}

\item{var}{the name of the variable}

\item{var_name}{what to call the variable in the error message}
}
\value{
vector of error messages
}
\description{
Utility function to ensure that \code{var} contains no missing values
}
