% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{check_names}
\alias{check_names}
\title{Checks if all variable names are present}
\usage{
check_names(df)
}
\arguments{
\item{df}{VIS data.frame}
}
\value{
vector of error messages
}
\description{
Checks if all variable names are present
}
