% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{check_reasoncode}
\alias{check_reasoncode}
\title{Checks that all ReasonCodes are correct}
\usage{
check_reasoncode(df)
}
\arguments{
\item{df}{VIS data.frame}
}
\value{
vector of error messages
}
\description{
\itemize{
\item Ensures that each ReasonCode is NA, "MSK", "NI", "INV", or "UNK"
}
}
