% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{check_register_hsaid}
\alias{check_register_hsaid}
\title{Checks if register HSAID is correct}
\usage{
check_register_hsaid(df)
}
\arguments{
\item{df}{VIS data.frame}
}
\value{
vector of error messages
}
\description{
\itemize{
\item Ensures that the HSAID in the instruction cell matches all other observed
register HSAIDs
\item Ensures that there is only one register HSAID
}
}
