% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{check_unit_hospital}
\alias{check_unit_hospital}
\title{Checks that all censored rows have values removed}
\usage{
check_unit_hospital(df)
}
\arguments{
\item{df}{VIS data.frame}
}
\value{
vector of error messages
}
\description{
\itemize{
\item Ensures that only one of UnitHSAID, HospitalHSAID, and HospitalOrganisationID is present
\item Ensures that only one of UnitName and HospitalName is present
}
}
