% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{check_value_rate}
\alias{check_value_rate}
\title{Checks if all rows have Value or Rate}
\usage{
check_value_rate(df)
}
\arguments{
\item{df}{VIS data.frame}
}
\value{
vector of error messages
}
\description{
\itemize{
\item Ensures that all rows that are not censored have Value or Rate
\item Ensures that all Rates have Numerator and Denominator
}
}
