% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{check_version}
\alias{check_version}
\title{Checks that Version is correct}
\usage{
check_version(df)
}
\arguments{
\item{df}{VIS data.frame}
}
\value{
vector of error messages
}
\description{
Checks that Version is correct
}
