% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_table_one.R
\name{classify_table_one_vars}
\alias{classify_table_one_vars}
\title{Classify \code{vars} into binary, categorical, or numeric}
\usage{
classify_table_one_vars(df, vars)
}
\arguments{
\item{df}{a data.frame}

\item{vars}{the variables to classify}
}
\value{
a character vector of the same length as \code{vars}, with one of
\code{"binary"}, \code{"categorical"}, or \code{"numeric"} for each entry
}
\description{
Classify \code{vars} into binary, categorical, or numeric
}
