% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_surv.R
\name{convert}
\alias{convert}
\title{Convert an object to a specified class}
\usage{
convert(x, cls)
}
\arguments{
\item{x}{object to convert}

\item{cls}{class to convert to}
}
\value{
x converted to specified class
}
\description{
Convert an object to a specified class
}
