% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_labels.R
\name{decode_data}
\alias{decode_data}
\title{Replace each value with its corresponding label}
\usage{
decode_data(
  data,
  labels = NULL,
  droplevels = TRUE,
  add_cols = FALSE,
  suffix = "_label",
  as_character = FALSE,
  missing_labels_na = TRUE,
  decode_names = FALSE,
  variable_names = NULL
)
}
\arguments{
\item{data}{Data frame with Stratum register data}

\item{labels}{Value labels for the same register}

\item{droplevels}{If \code{TRUE}, factor levels not present in data will be
excluded. If \code{FALSE} no levels are dropped. If character
then the specified variables will have its levels dropped.}

\item{add_cols}{Add additional columns with the labels}

\item{suffix}{suffix to add to added columns}

\item{as_character}{If \code{TRUE} variables
will be set to characters. If \code{FALSE} variables will be factors.}

\item{missing_labels_na}{If \code{TRUE} values with no label will be set to \code{NA}}

\item{decode_names}{If \code{TRUE} column names will be decode aswel}

\item{variable_names}{Names for columns}
}
\value{
data.frame with character values instead
of numerical values.
}
\description{
Returns data with descriptive data points
instead of codes.
}
\examples{
\dontrun{
factor_data <- decode_vars(data, factor_levels)
}

}
