% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{extract_function_call}
\alias{extract_function_call}
\title{Extract r function call from RC-statisticts API-URL}
\usage{
extract_function_call(api_url)
}
\arguments{
\item{api_url}{The URL of a function call to a RC-statistics endpoint.}
}
\value{
a character vector of a r function call to a statistics API-endpoint
}
\description{
Extract r function call from RC-statisticts API-URL
}
\examples{
extract_function_call(
api_url = "https://sep.registercentrum.se/stratum/api/statistics/sep/sep-jumbo-antalreg?apikey=bK3H9bwaG4o="
)
}
