% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_table_one.R
\name{extract_table_one_mean}
\alias{extract_table_one_mean}
\title{Extract mean variables from the result of \code{\link[=get_aggregate_value]{get_aggregate_value()}}}
\usage{
extract_table_one_mean(res, mean_vars, group_cols, digits)
}
\arguments{
\item{res}{the result of \code{\link[=get_aggregate_value]{get_aggregate_value()}}}

\item{mean_vars}{the numeric variables to summarize as mean (SD)}

\item{group_cols}{the columns to group by}

\item{digits}{the number of decimals used when formatting statistics}
}
\value{
a long data.frame with one row per variable
}
\description{
Extract mean variables from the result of \code{\link[=get_aggregate_value]{get_aggregate_value()}}
}
