% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_table_one.R
\name{extract_table_one_prop}
\alias{extract_table_one_prop}
\title{Extract binary variables from the result of \code{\link[=get_aggregate_value]{get_aggregate_value()}}}
\usage{
extract_table_one_prop(res, binary_vars, group_cols, digits)
}
\arguments{
\item{res}{the result of \code{\link[=get_aggregate_value]{get_aggregate_value()}}}

\item{binary_vars}{the binary variables to extract}

\item{group_cols}{the columns to group by}

\item{digits}{the number of decimals used when formatting statistics}
}
\value{
a long data.frame with one row per variable
}
\description{
Extract binary variables from the result of \code{\link[=get_aggregate_value]{get_aggregate_value()}}
}
