% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_table_one.R
\name{extract_table_one_prop_count}
\alias{extract_table_one_prop_count}
\title{Extract categorical variables from the result of \code{\link[=get_aggregate_value]{get_aggregate_value()}}}
\usage{
extract_table_one_prop_count(res, categorical_vars, group_cols, digits)
}
\arguments{
\item{res}{the result of \code{\link[=get_aggregate_value]{get_aggregate_value()}}}

\item{categorical_vars}{the categorical variables to extract}

\item{group_cols}{the columns to group by}

\item{digits}{the number of decimals used when formatting statistics}
}
\value{
a long data.frame with one row per variable/category
}
\description{
Extract categorical variables from the result of \code{\link[=get_aggregate_value]{get_aggregate_value()}}
}
