% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_table_one.R
\name{format_table_one_mean}
\alias{format_table_one_mean}
\title{Format a mean/SD pair as \code{"mean (SD)"}}
\usage{
format_table_one_mean(mean, sd, digits)
}
\arguments{
\item{mean}{a numeric vector of means}

\item{sd}{a numeric vector of standard deviations}

\item{digits}{the number of decimals used when formatting \code{mean}/\code{sd}}
}
\value{
a character vector
}
\description{
Format a mean/SD pair as \code{"mean (SD)"}
}
