% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_table_one.R
\name{format_table_one_median}
\alias{format_table_one_median}
\title{Format a median/range triple as \code{"median (5\%-95\%)"}}
\usage{
format_table_one_median(median, quant_5, quant_95, digits)
}
\arguments{
\item{median}{a numeric vector of medians}

\item{quant_5}{a numeric vector of 5th percentiles}

\item{quant_95}{a numeric vector of 95th percentiles}

\item{digits}{the number of decimals used when formatting the statistics}
}
\value{
a character vector
}
\description{
Format a median/range triple as \code{"median (5\%-95\%)"}
}
