% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_table_one.R
\name{format_table_one_prop}
\alias{format_table_one_prop}
\title{Format a count/proportion pair as \code{"N (P\%)"}}
\usage{
format_table_one_prop(n, prop, digits)
}
\arguments{
\item{n}{a numeric vector of counts}

\item{prop}{a numeric vector of proportions}

\item{digits}{the number of decimals used when formatting \code{prop}}
}
\value{
a character vector
}
\description{
Format a count/proportion pair as \code{"N (P\%)"}
}
