% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gender}
\alias{gender}
\title{Derives gender from Swedish social security number}
\usage{
gender(x)
}
\arguments{
\item{x}{Swedish social security number}
}
\value{
1 if male, 2 if female
}
\description{
Derives gender from Swedish social security number
}
