% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_statistics.R
\name{get_group_combinations}
\alias{get_group_combinations}
\title{Get a list of all combinations of columns defined in \code{group_cols}}
\usage{
get_group_combinations(group_cols, marginal_cols)
}
\arguments{
\item{group_cols}{the columns to group by}

\item{marginal_cols}{the columns for which to create marginal summaries}
}
\value{
a list of combinations
}
\description{
Also finds which columns to always group by and adds these to the
combinations.
}
