% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_vis.R
\name{get_measure_config}
\alias{get_measure_config}
\title{Gets a minimal representation of a measure config for a specific measure ID}
\usage{
get_measure_config(measure_id)
}
\arguments{
\item{measure_id}{the measure ID to search for}
}
\value{
a list containing the version, measure frequencies, organizational
levels, title, and status
}
\description{
Gets a minimal representation of a measure config for a specific measure ID
}
