% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_labels.R
\name{get_sos_vl}
\alias{get_sos_vl}
\title{Funktion för att hämta value labels från Socialstyrelsen}
\usage{
get_sos_vl(register = NULL)
}
\arguments{
\item{register}{Register på Socialstyrelsen}
}
\description{
Funktion för att hämta value labels från Socialstyrelsen
}
