% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_vis.R
\name{get_submeasures}
\alias{get_submeasures}
\title{Gets the specified measure submeasures from a measure config}
\usage{
get_submeasures(measure_config)
}
\arguments{
\item{measure_config}{the measure config from which to extract the
submeasures, the result of \code{\link[=search_measure_id]{search_measure_id()}}}
}
\value{
list of gender codes and submeasure IDs
}
\description{
Gets the specified measure submeasures from a measure config
}
