% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_surv.R
\name{get_surv_value}
\alias{get_surv_value}
\title{Gets the estimated survival at a specific point in time}
\usage{
get_surv_value(
  df,
  time,
  time_col = "time",
  event_col = "status",
  group_cols = NULL,
  marginal_cols,
  obfuscate_data = FALSE,
  add_reason_col = TRUE,
  censored_value = NA,
  censored_limit = 15,
  estimate = "survival",
  include_ci = FALSE
)
}
\arguments{
\item{df}{data.frame}

\item{time}{point in time (in days)}

\item{time_col}{the name of the time column}

\item{event_col}{the name of the event column}

\item{group_cols}{the columns to group by}

\item{marginal_cols}{the columns to add marginal groups for}

\item{obfuscate_data}{whether or not to obfuscate data}

\item{add_reason_col}{whether or not to add the reason for data being
censored}

\item{censored_value}{the value to use when observations are censored}

\item{censored_limit}{lower limir for at risk before censoring}

\item{estimate}{what the estimate should be, either "survival" or "event"}

\item{include_ci}{should confidence intervals be included in the final
data.frame}
}
\value{
estimated survival at the requested point in time for each group
}
\description{
Gets the estimated survival at a specific point in time
}
