% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{get_vis_names}
\alias{get_vis_names}
\title{Get a vector of expect variable names for a VIS file}
\usage{
get_vis_names()
}
\value{
a vector of variable names
}
\description{
Get a vector of expect variable names for a VIS file
}
