% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_statistik.R
\name{group_means}
\alias{group_means}
\title{Calculate n, means and sd by group}
\usage{
group_means(data, ..., vars = names(data))
}
\arguments{
\item{data}{A data.frame or tibble}

\item{...}{Variables to group on by}

\item{vars}{Variables to calculate means and sd on.
Defaults to all vars in data.}
}
\description{
Calculate n, means and sd by group
}
