% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_male}
\alias{is_male}
\title{Derives gender from Swedish social security number}
\usage{
is_male(x)
}
\arguments{
\item{x}{Swedish social security number}
}
\value{
boolean
}
\description{
Derives gender from Swedish social security number
}
