% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_table_one.R
\name{is_table_one_binary}
\alias{is_table_one_binary}
\title{Check if a vector should be treated as binary}
\usage{
is_table_one_binary(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a single logical
}
\description{
\code{logical} vectors, and numeric vectors only containing \code{0}/\code{1} (and
\code{NA}), are treated as binary.
}
