% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locf.R
\name{locfdt}
\alias{locfdt}
\title{locfdt}
\usage{
locfdt(
  dt,
  vars,
  groupby = "id",
  orderby = "date",
  slice = FALSE,
  return_tibble = FALSE,
  window_size = NULL,
  window_type = NULL
)
}
\arguments{
\item{dt}{data.frame or data.table to do LOCF imputation on}

\item{vars}{variables to impute with LOCF}

\item{groupby}{Variable to group on}

\item{orderby}{Variable to sort on}

\item{slice}{If \code{TRUE} will select the last row in each group of \code{groupby}}

\item{return_tibble}{whether to return a data.table or a tibble}

\item{window_size}{The maximum time the last observation can be
carried forward. If \code{n_months} is \code{NULL} LOCF will be carried out
on all observations like normal.}

\item{window_type}{\code{years}, \code{months} or \code{days}, deafult is \code{days}.
With \code{months} and \code{years} the \code{window_size} is simply multiplied by a
factor of 31 or 365.25 respectively.}
}
\value{
data.frame with LOCF imputation
}
\description{
DT Last Observation Carried Forward with
}
\details{
Returns data.frame with LOCF imputation using
data.table. Faster than \code{RCStat::locf} for large
datasets, especially when imputing on multiple variables
}
\examples{
\dontrun{
df <- locfdt(df, vars)
}

}
