% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_table_one.R
\name{lump_table_one_categories}
\alias{lump_table_one_categories}
\title{Lump the smallest categories of a categorical variable into \code{"Other"}}
\usage{
lump_table_one_categories(x, max_categories, name, other_level = "Other")
}
\arguments{
\item{x}{a vector}

\item{max_categories}{the maximum number of categories to keep}

\item{name}{the name of the variable, used in the warning message}

\item{other_level}{value of the level used for "other" values, passed to
\code{\link[forcats:fct_lump_n]{forcats::fct_lump_n()}}}
}
\value{
\code{x} as a factor, with the smallest categories lumped into
\code{"Other"} if \code{x} has more than \code{max_categories} categories
}
\description{
Lump the smallest categories of a categorical variable into \code{"Other"}
}
