% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteSAS.R
\name{make_SAS_formats}
\alias{make_SAS_formats}
\title{make_SAS_formats}
\usage{
make_SAS_formats(varnames)
}
\arguments{
\item{varnames}{names of data.frame}
}
\value{
format for each variable
}
\description{
Define format for SAS code
}
\examples{
\dontrun{names <- make_SAS_formats(vars)}
}
