% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteSAS.R
\name{make_SAS_names}
\alias{make_SAS_names}
\title{make_SAS_names}
\usage{
make_SAS_names(varnames, validvarname = c("V7", "V6"))
}
\arguments{
\item{varnames}{names of data.frame}

\item{validvarname}{nmax}
}
\value{
modified variable names
}
\description{
Modify column names to be valid in SAS
}
\examples{
\dontrun{names <- make_SAS_names(vars, vars)}
}
