% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_surv.R
\name{obfuscate_surv}
\alias{obfuscate_surv}
\title{Obfuscate result from \code{\link[=get_surv_value]{get_surv_value()}}}
\usage{
obfuscate_surv(
  df,
  add_reason_col = TRUE,
  censored_value = NA,
  censored_limit = 15
)
}
\arguments{
\item{df}{data.frame}

\item{add_reason_col}{add obfuscated reason}

\item{censored_value}{the value to use when censoring observations}

\item{censored_limit}{lower limir for at risk before censoring}
}
\value{
data.frame
}
\description{
Obfuscate result from \code{\link[=get_surv_value]{get_surv_value()}}
}
