% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_statistik.R
\name{proportion_missing}
\alias{proportion_missing}
\title{Calculates proportion of missing data}
\usage{
proportion_missing(data, ..., vars = names(data))
}
\arguments{
\item{data}{data.frame or tibble}

\item{...}{grouping to apply before calculation}

\item{vars}{vars to calculate proportion of missing
data on. Defaults to all except grouping vars.}
}
\description{
Calculates n missing and proportion missing across
columns and by groups specified.
}
