% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pseudonymize_data}
\alias{pseudonymize_data}
\title{Pseudonymizes a data.frame}
\usage{
pseudonymize_data(
  df,
  pseudonimzed_var = "lopnr",
  subject_key = "SubjectKey",
  remove_subject_key = TRUE,
  save_key = TRUE,
  key_file_dir
)
}
\arguments{
\item{df}{the data.frame to pseudonymize}

\item{pseudonimzed_var}{the name of the variable to store the key in}

\item{subject_key}{the name of the variable containing the subject identifier}

\item{remove_subject_key}{whether to remove the variable after adding the keys}

\item{save_key}{whether to save the key as an Excel-file}

\item{key_file_dir}{the directory in which to save the key}
}
\value{
the data.frame passed to \code{df} with a pseudonymized variable added
}
\description{
Pseudonymizes a data.frame
}
