% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_mail.R
\name{rc_mail}
\alias{rc_mail}
\title{Function to send e-mail from Registercentrum
office 365 mail.}
\usage{
rc_mail(to, from, subject, body, password)
}
\arguments{
\item{to}{e-mail to send to}

\item{from}{Name of sender, included in signature}

\item{subject}{Subject of message}

\item{body}{Text to include in body of message.
Please note that any line breaks must be done in
windows style to be accepted by mail server (eg.
use \\r\\n to break a line.).}

\item{password}{The password to the statistik\@registercentrum.se email}
}
\description{
Function to send e-mail from Registercentrum
office 365 mail.
}
\examples{
\dontrun{
rc_mail(
  to = "rikard.isaksson@vgregion.se",
  from = "Rikard Isaksson",
  subject = "Hej test",
  body = "Test!")

}
}
