% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rojande.R
\name{reason_col}
\alias{reason_col}
\title{Internal function to add obfuscated reason column}
\usage{
reason_col(
  data,
  liberal_obfuscation = TRUE,
  count_var = "n",
  total_var = "total",
  prop_var = "prop",
  statistics_vars = NULL
)
}
\arguments{
\item{data}{Data to be obfuscated}

\item{liberal_obfuscation}{Whether or not to use liberal obfuscation}

\item{count_var}{The column containing the numerator}

\item{total_var}{The column containing the denominator}

\item{prop_var}{The column containing the proportion as determined by
\code{count_var} and \code{total_var}}

\item{statistics_vars}{Columns containing statistics such as mean, sd, etc.}
}
\description{
Internal function to add obfuscated reason column
}
