% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{return_errors}
\alias{return_errors}
\title{Format errors before return}
\usage{
return_errors(errors, as_data_frame)
}
\arguments{
\item{errors}{the vector of errors}

\item{as_data_frame}{logical indicating whether to return a data.frame or vector}
}
\value{
a data.frame or vector of errors
}
\description{
Format errors before return
}
