% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_mean.R
\name{rolling_mean}
\alias{rolling_mean}
\title{Rolling average over vector}
\usage{
rolling_mean(x, n, align = c("left", "center", "right")[3], na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector to compute rolling average on}

\item{n}{size of window}

\item{align}{start window at the left, right or middle}

\item{na.rm}{remove NA when computing mean (consider them to be 0)}
}
\description{
Rolling average over vector
}
