% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_statistics.R
\name{rollup_prop_count_one}
\alias{rollup_prop_count_one}
\title{Roll finest-grain prop_count counts up to one group_cols combination}
\usage{
rollup_prop_count_one(
  finest_counts,
  comb,
  var_name,
  include_missing,
  obfuscate_data,
  censored_value,
  add_reason_col
)
}
\arguments{
\item{finest_counts}{output of \code{\link[=build_finest_prop_count]{build_finest_prop_count()}}}

\item{comb}{the subset of columns to group by for this combination}

\item{var_name}{the prop_count column name, used to prefix output columns}

\item{include_missing}{whether NA values of the tabulated var count as
a category}

\item{obfuscate_data}{whether counts should be obfuscated}

\item{censored_value}{the value to replace censored counts with}

\item{add_reason_col}{whether to add an obfuscation-reason column}
}
\value{
a wide tibble keyed by \code{comb}'s columns
}
\description{
Roll finest-grain prop_count counts up to one group_cols combination
}
