% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roundc}
\alias{roundc}
\alias{round_half_up}
\title{Rounds numbers with halves being rounded up}
\usage{
roundc(x, digits = 0)

round_half_up(x, digits = 0)
}
\arguments{
\item{x}{vector of values}

\item{digits}{number of decimals to show after rounding}
}
\value{
vector of rounded values
}
\description{
Rounds numbers with halves being rounded up
}
\section{Functions}{
\itemize{
\item \code{roundc()}: deprecated Rcpp implementation

}}
