% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rojande.R
\name{round_to_tens}
\alias{round_to_tens}
\title{Round integer to nearest 10}
\usage{
round_to_tens(n)
}
\arguments{
\item{n}{integer, normally a frequency}
}
\description{
5's rounds to nearest even, i.e. 15 rounds to 20
and 25 rounds to 20. See ?round.
}
