% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rojande.R
\name{round_to_y}
\alias{round_to_y}
\title{Round number to nearest y}
\usage{
round_to_y(x, y = 0.05)
}
\arguments{
\item{x}{The number to round}

\item{y}{The number to round to}
}
\description{
Round number to nearest y
}
\examples{
round_to_y(x = 0.12, y = 0.05)
}
