% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roundc}
\alias{roundc}
\title{Rounding with ties-away-from-zero}
\usage{
roundc(x, digits = 0L)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{number of digits to round to}
}
\value{
rounded vector
}
\description{
Rounding with ties-away-from-zero
}
