% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rojande.R
\name{rounded_ci_p}
\alias{rounded_ci_p}
\title{Get rounded ci for proportion}
\usage{
rounded_ci_p(p_hat, n, alpha = 0.05)
}
\arguments{
\item{p_hat}{estimate of proportion}

\item{n}{sample size}

\item{alpha}{alpha level}
}
\description{
Get rounded ci for proportion
}
