% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_vis.R
\name{search_measure_id}
\alias{search_measure_id}
\title{Sends a call to the Kvalitetsindikatorkatalog API to retrieve information
about the specified measure ID}
\usage{
search_measure_id(measure_id)
}
\arguments{
\item{measure_id}{the measure ID to search for}
}
\value{
the first result of the API call
}
\description{
Sends a call to the Kvalitetsindikatorkatalog API to retrieve information
about the specified measure ID
}
