% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_labels.R
\name{set_factors}
\alias{set_factors}
\title{set_factors}
\usage{
set_factors(df, labels, droplevels = TRUE)
}
\arguments{
\item{df}{Data frame with Stratum register data}

\item{labels}{Value labels for the same register}

\item{droplevels}{If `TRUE`, factor levels not present in data will be
excluded. If `FALSE` no levels are dropped. If character
then the specified variables will have its levels dropped.}
}
\value{
data.frame with character values instead
        of numerical values.
}
\description{
Replace each value with its corresponding label
}
\details{
Returns data frame with descriptive data points
instead of numerical codes.
}
\examples{
\dontrun{
factor_data <- set_factors(data, factor_levels)
}

}
