% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_labels.R
\name{set_value_labels}
\alias{set_value_labels}
\title{set_value_labels}
\usage{
set_value_labels(df, labels)
}
\arguments{
\item{df}{Data frame with Stratum register data}

\item{labels}{Value labels for the same register/form}
}
\value{
data.frame with map attribute labels on values.
}
\description{
Assign each value with its corresponding label
}
\details{
Returns data frame with descriptive labels for data points
}
\examples{
\dontrun{
df <- set_value_labels(df, lab_val)
}
}
