% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_labels.R
\name{SetValueLabels}
\alias{SetValueLabels}
\title{SetValueLabels}
\usage{
SetValueLabels(df, labels)
}
\arguments{
\item{df}{Data frame with Stratum register data}

\item{labels}{Value labels for the same register/form}
}
\value{
data.frame with map attribute labels on values.
}
\description{
Assign each value with its corresponding label
}
\details{
Returns data frame with descriptive labels for data points
}
\examples{
\dontrun{
df <- GetRegisterData(1012, RunDeep = TRUE)
lab_val <- GetValueLabels(1012)
df <- SetValueLabels(df, lab_val)
}
}
