% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_mail.R
\name{sh_mail}
\alias{sh_mail}
\title{Function to generate bash-script that sends
e-mail with attachment from Registercentrum
office 365 mail.}
\usage{
sh_mail(to, subject, password, file = NULL, run_bash = T)
}
\arguments{
\item{to}{e-mail to send to}

\item{subject}{Subject of message}

\item{password}{The password to the statistik@registercentrum.se email}

\item{file}{file to include. Currently supports
one file of type csv, xlsx, pdf, zip, doc or docx.}

\item{run_bash}{Whether or not to run and then delete the created bash script}
}
\description{
Function to generate bash-script that sends
e-mail with attachment from Registercentrum
office 365 mail.
}
\examples{
\dontrun{
sh_mail(
  to = "rikard.isaksson@vgregion.se",
  subject = "Hej test",
  file = "data/test.csv")

sh_mail(
  to = c("alexander.thoren@vgregion.se",
 "oskar.e.johansson@vgregion.se"),
  subject = "test",
  password = ...,
  file = "test.xlsx")
}
}
