% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_pnr}
\alias{valid_pnr}
\title{Checks if the Swedish social security number is correctly formatted and
has the correct control number.}
\usage{
valid_pnr(pnr, handle_invalid = TRUE)
}
\arguments{
\item{pnr}{Swedish social security number}

\item{handle_invalid}{How to process social security numbers with invalid
formatting. If TRUE returns FALSE, if FALSE returns NA}
}
\value{
Boolean indicating correct (TRUE) or incorrect (FALSE) social
security number
}
\description{
Checks if the Swedish social security number is correctly formatted and
has the correct control number.
}
