% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_pnr_format}
\alias{valid_pnr_format}
\title{Checks if the Swedish social security number is in the expected stratum
format "XXXXXXXX-XXXX"}
\usage{
valid_pnr_format(pnr)
}
\arguments{
\item{pnr}{Swedish social security number}
}
\value{
Boolean indicating correct (TRUE) or incorrect (FALSE) format
}
\description{
Checks if the Swedish social security number is in the expected stratum
format "XXXXXXXX-XXXX"
}
