% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vis.R
\name{validate_vis}
\alias{validate_vis}
\title{Validates a VIS data.frame}
\usage{
validate_vis(
  df,
  overall_rate = 0.25,
  measure_rate = 0.75,
  as_data_frame = FALSE
)
}
\arguments{
\item{df}{VIS data.frame}

\item{overall_rate}{the overall censor rate to warn at}

\item{measure_rate}{the per-MeasureID censor rate to warn at}

\item{as_data_frame}{whether to return the errors in a data.frame}
}
\value{
a character vector of error messages
}
\description{
Checks if:
\itemize{
\item all variables are present
\item the instruction on the first line is correct
\item register HSAIDs are correct
\item dates are correct
\item Value and Rate are correct
\item confidence intervals are correct
\item values have not been censored to an unusually high degree
\item county and unit variables are correct
\item no rows are duplicated
}
}
\details{
If any variables are found to be missing the function exists early because
then we can not be sure remaining validations will be possible.
}
