% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_labels.R
\name{var_to_char}
\alias{var_to_char}
\title{Convert variable to character}
\usage{
var_to_char(x, labels, name = NULL, missing_labels_na = TRUE)
}
\arguments{
\item{x}{Vector with codes to decode}

\item{labels}{data.frame with columns ValueCode and ValueName to use when
decoding/translating.}

\item{name}{name of input variable (optional)}

\item{missing_labels_na}{If \code{TRUE} values with no label will be set to \code{NA}}
}
\value{
character vector
}
\description{
Used to extract texts from a variable containing numeric values.
}
\examples{

x <- sample(1:3, size = 10, replace = TRUE)

labels <- data.frame(
  ValueCode = 1:3,
  ValueName = c("Volvo", "Saab", "Opel")
)
var_to_char(x, labels, name = "cars")

}
