% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_labels.R
\name{var_to_factor}
\alias{var_to_factor}
\title{Convert vector to factor using supplied labels}
\usage{
var_to_factor(
  x,
  labels,
  name = NULL,
  droplevels = TRUE,
  missing_labels_na = TRUE
)
}
\arguments{
\item{x}{Vector with codes to decode}

\item{labels}{data.frame with columns ValueCode and ValueName to use
when decoding.}

\item{name}{name of input variable (optional). Used in the warning message
only to print the name of variable with missing labels.}

\item{droplevels}{If \code{TRUE}, factor levels not present in data will be
excluded.}

\item{missing_labels_na}{If \code{TRUE} values with no label will be set to \code{NA}}
}
\value{
Factor
}
\description{
Convert vector to factor using supplied labels
}
\examples{

x <- sample(1:3, size = 10, replace = TRUE)

labels <- data.frame(
  ValueCode = 1:3,
  ValueName = c("Volvo", "Saab", "Opel")
)
var_to_factor(x, labels, name = "cars")
}
