% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteSAS.R
\name{WriteSAS}
\alias{WriteSAS}
\title{WriteSAS}
\usage{
WriteSAS(
  df,
  datafile,
  codefile,
  dataname = "rdata",
  validvarname = c("V7", "V6"),
  libpath = getwd()
)
}
\arguments{
\item{df}{names of data.frame}

\item{datafile}{name of output data}

\item{codefile}{name of generated SAS code}

\item{dataname}{name of sas-dataset}

\item{validvarname}{eeeh}

\item{libpath}{path for lib in SAS code}
}
\value{
format for each variable
}
\description{
Define format for SAS code
}
\examples{
\dontrun{
names <-
  WriteSAS(
    df, datafile = 'out_data.csv',
    codefile = 'sas_code.sas',
  )
}
}
