% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zip_dir_with_pass}
\alias{zip_dir_with_pass}
\title{Zip directory with password}
\usage{
zip_dir_with_pass(
  directory = "output",
  pass = random_password(),
  file = "output",
  overwrite = TRUE,
  sink_password = TRUE
)
}
\arguments{
\item{directory}{directory to zip. Defaults to \code{output}}

\item{pass}{password to use. Defaults to \verb{[random_password()]}. May only
contain \verb{a–z}, \verb{A–Z}, \verb{0–9}, and \code{!}, \code{@}, \verb{#}, \code{$}, \code{.}, \code{+}, \verb{\%}, \code{&}, \code{/},
\code{(}, \verb{)}, \verb{?}.}

\item{file}{name of zip-file}

\item{overwrite}{Overwrite zip-file if it exists}

\item{sink_password}{whether or not to save the password in a file called
password.txt in the current working directory}
}
\description{
zips chosen directory with password protection. Zip-file and
txt-file with the password are placed in working directory
}
\examples{
\dontrun{
zip_dir_with_pass()
}
}
