% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_vis.R
\name{add_groups_long}
\alias{add_groups_long}
\alias{add_groups_long_date}
\alias{add_groups_long_organization}
\title{Adds appropriate groups in long format}
\usage{
add_groups_long(df, cfg, date_var, unit_var, county_var)

add_groups_long_date(df, date_var, type)

add_groups_long_organization(df, unit_var, county_var, type)
}
\arguments{
\item{df}{a data.frame}

\item{cfg}{the measure config, the result of [get_measure_config()]}

\item{date_var}{the name of the date variable}

\item{unit_var}{the name of the unit variable}

\item{county_var}{the name of county variable}

\item{type}{the type of date or organizational level to be added}
}
\value{
a data.frame with extra rows added
}
\description{
Adds appropriate groups in long format
}
\section{Functions}{
\itemize{
\item \code{add_groups_long_date()}: adds long date groups

\item \code{add_groups_long_organization()}: adds long organization (unit and county) groups

}}
