% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_statistics.R
\name{api_url}
\alias{api_url}
\alias{api_statistics}
\title{Get data from statistics API}
\usage{
api_url(
  register,
  fun,
  arguments = list(),
  api_key = "MpuYxfbtp5I=",
  dev = FALSE,
  api_url = "prod"
)

api_statistics(
  register,
  fun,
  arguments = list(),
  api_key = "MpuYxfbtp5I=",
  dev = FALSE,
  verbose = FALSE,
  return_tibble = TRUE,
  api_url = "prod"
)
}
\arguments{
\item{register}{register short name}

\item{fun}{script name}

\item{arguments}{Arguments to include as named list}

\item{api_key}{api-key to use}

\item{dev}{if newr should be used}

\item{api_url}{Should be `"prod"` for calling the production Stratum API or
`"local"` to call a local instance of `rcplumber`. All other values will
interpreted as the prefix for the URL of the API call. Can also be controlled
by the option `rcstat.local_plumber`}

\item{verbose}{print query sent}

\item{return_tibble}{whether or not to coerce data to a tibble using
tibble::as_tibble. Note that this does not work when the data is a list.}
}
\description{
Get data from statistics API
}
\section{Functions}{
\itemize{
\item \code{api_url()}: gets the url for the API call

}}
