% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_statistics.R
\name{api_statistics}
\alias{api_statistics}
\title{Get data from statistics API}
\usage{
api_statistics(
  register,
  fun,
  arguments = list(),
  api_key = "MpuYxfbtp5I=",
  dev = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{register}{register short name}

\item{fun}{script name}

\item{arguments}{Arguments to include as named list}

\item{api_key}{api-key to use}

\item{dev}{if newr should be used}

\item{verbose}{print query sent}
}
\description{
Get data from statistics API
}
