% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_statistics.R
\name{count_prop_wide}
\alias{count_prop_wide}
\title{Summarise count into wide format}
\usage{
count_prop_wide(x, include_missing = FALSE, obfuscate_data, censored_value)
}
\arguments{
\item{x}{A vector, numeric or character}

\item{include_missing}{logical indicating whether or not to include
NA values}

\item{obfuscate_data}{logical indicating whether or not to obfuscate data}

\item{censored_value}{the value to replace censored proportions with}
}
\description{
Summarise a vector into columns of counts for each occurrence and each counts
associated proportion
}
