% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_labels.R
\name{decode_names}
\alias{decode_names}
\title{Decode variabel names}
\usage{
decode_names(df, labels = NULL)
}
\arguments{
\item{df}{data.frame}

\item{labels}{data frame containing column names}
}
\value{
data frame with description as column names
}
\description{
Replaces the names of the columns with description of column
}
