% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{eq5d_3l_index_tto}
\alias{eq5d_3l_index_tto}
\title{Calculate Swedish EQ5D-index based on TTO from EQ5D-3-level data}
\usage{
eq5d_3l_index_tto(mobility, selfcare, usual, pain, anxiety)
}
\arguments{
\item{mobility, selfcare, usual, pain, anxiety}{numeric
scale values for each dimension}
}
\value{
numeric vector with calulated index-values
}
\description{
Calculate Swedish EQ5D-index based on TTO from EQ5D-3-level data
}
\examples{
eq5d_3l_index_tto(1, 1, 1, 1, 1) # 0.969
eq5d_3l_index_tto(1, 2, 2, 3, 3) # 0.499
eq5d_3l_index_tto(2, 1, 3, 1, 1) # 0.724

# Works with vectors as well
eq5d_3l_index_tto(rep(1, 3), rep(1, 3), rep(1, 3), rep(1, 3), rep(1, 3))
}
\references{
[Burström et al. 2014, TTO Model 4, Table 3](https://dx.doi.org/10.1007%2Fs11136-013-0496-4)
}
\seealso{
Other eq5d: 
\code{\link{eq5d_3l_index_tto_old}()},
\code{\link{eq5d_5l_index_tto}()},
\code{\link{eq5d_5l_index_vas}()},
\code{\link{eq5d_index}()}
}
\concept{eq5d}
