% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{eq5d_3l_index_tto_old}
\alias{eq5d_3l_index_tto_old}
\title{Convert EQ5D-3L-values stored in combined character vectors to Swedish
EQ5D-index based on TTO}
\usage{
eq5d_3l_index_tto_old(x, split = ",")
}
\arguments{
\item{x}{character vector with comma-separated vaslues for each dimension.}

\item{split}{
    character vector (or object which can be coerced to such)
    containing \link[base]{regular expression}(s) (unless \code{fixed = TRUE})
    to use for splitting.  If empty matches occur, in particular if
    \code{split} has length 0, \code{x} is split into single characters.
    If \code{split} has length greater than 1, it is re-cycled along
    \code{x}.
  }
}
\value{
numeric vector with EQ5D-indices based on TTO
}
\description{
Convert EQ5D-3L-values stored in combined character vectors to Swedish
EQ5D-index based on TTO
}
\examples{
eq5d_3l_index_tto_old(c("2,2,2,2,3", NA, "2,2,2,3,3", "2,2,2,2,2"))
eq5d_3l_index_tto_old(c("22223", NA, "22233", "22222"), split = "")
}
\seealso{
\code{\link[=eq5d_3l_index_tto]{eq5d_3l_index_tto()}}

Other eq5d: 
\code{\link{eq5d_3l_index_tto}()},
\code{\link{eq5d_5l_index_tto}()},
\code{\link{eq5d_5l_index_vas}()},
\code{\link{eq5d_index}()}
}
\concept{eq5d}
