% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{eq5d_5l_index_tto}
\alias{eq5d_5l_index_tto}
\title{Calculate Swedish EQ5D-index based on TTO from EQ5D-5-level data}
\usage{
eq5d_5l_index_tto(mobility, selfcare, usual, pain, anxiety)
}
\arguments{
\item{mobility, selfcare, usual, pain, anxiety}{numeric
scale values for each dimension}
}
\value{
numeric vector with EQ5D-indices
}
\description{
Calculate Swedish EQ5D-index based on TTO from EQ5D-5-level data
}
\examples{
eq5d_5l_index_tto(3, 4, 5, 4, 3) # 0.503
eq5d_5l_index_tto(rep(3, 10), rep(4, 10), rep(5, 10), rep(4, 10), rep(3, 10))
}
\references{
\href{https://link.springer.com/article/10.1007/s40273-020-00905-7/tables/9}{Burström, model 5 TTO}
}
\seealso{
Other eq5d: 
\code{\link{eq5d_3l_index_tto}()},
\code{\link{eq5d_3l_index_tto_old}()},
\code{\link{eq5d_5l_index_vas}()},
\code{\link{eq5d_index}()}
}
\concept{eq5d}
