% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{eq5d_index}
\alias{eq5d_index}
\title{Calculate Swedish EQ5D-indices based on TTO}
\usage{
eq5d_index(..., levels, type, old = FALSE, split = ",")
}
\arguments{
\item{...}{arguments passed to underlying methods.
Either \verb{mobility,selfcare,usual,pain,anxiety}
to \code{\link[=eq5d_3l_index_tto]{eq5d_3l_index_tto()}} or \code{\link[=eq5d_5l_index_tto]{eq5d_5l_index_tto()}}
or \code{x} to \code{\link[=eq5d_3l_index_tto_old]{eq5d_3l_index_tto_old()}}}

\item{levels}{numeric: 3 or 5 for the number of levels used}

\item{type}{character: "TTO" or "VAS"}

\item{old}{logic: should \code{\link[=eq5d_3l_index_tto_old]{eq5d_3l_index_tto_old()}} be used?}

\item{split}{
    character vector (or object which can be coerced to such)
    containing \link[base]{regular expression}(s) (unless \code{fixed = TRUE})
    to use for splitting.  If empty matches occur, in particular if
    \code{split} has length 0, \code{x} is split into single characters.
    If \code{split} has length greater than 1, it is re-cycled along
    \code{x}.
  }
}
\value{
numeric vector with calculated indices
}
\description{
The general purpose eq5d-package (available on CRAN) might be used for most
calculations of EQ5D-index values. Some combinations based on Swedish value
sets are, however, not (yet) available there.
}
\examples{
eq5d_index(1, 1, 1, 1, 1, levels = 3, type = "TTO")
eq5d_index(c("1,1,1,1,1"), levels = 3, type = "TTO", old = TRUE)
eq5d_index(1, 1, 1, 1, 1, levels = 5, type = "TTO")
eq5d_index(3, 4, 5, 4, 3, levels = 5, type = "VAS")
}
\seealso{
https://github.com/fragla/eq5d

Other eq5d: 
\code{\link{eq5d_3l_index_tto}()},
\code{\link{eq5d_3l_index_tto_old}()},
\code{\link{eq5d_5l_index_tto}()},
\code{\link{eq5d_5l_index_vas}()}
}
\concept{eq5d}
