% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_statistics.R
\name{get_aggregate_value}
\alias{get_aggregate_value}
\title{Create aggregate data across all combinations, including totals.
Can specify if proportion or average should be used.}
\usage{
get_aggregate_value(
  df,
  group_cols = NULL,
  vars = NULL,
  include_missing = TRUE,
  obfuscate_data = FALSE,
  censored_value = 0,
  pivot_prop_count = FALSE,
  distinct_cols = NULL,
  arrange_by = NULL,
  marginal_cols,
  add_reason_col = FALSE
)
}
\arguments{
\item{df}{A data frame.}

\item{group_cols}{The columns to group by.}

\item{vars}{Variables to be used to calculate the proportion from.}

\item{include_missing}{If missing values should be included in the total}

\item{obfuscate_data}{If data should be obfuscated}

\item{censored_value}{What value to replace censored values, used as argument in obfuscate_data}

\item{pivot_prop_count}{whether to pivot the resulting data.frame into long format}

\item{distinct_cols}{a set of columns that are used in
\code{\link[dplyr:distinct]{dplyr::distinct()}}. Should be disjoint from \code{group_cols} to prevent
misinterpretation of results.}

\item{arrange_by}{a column used in \code{\link[dplyr:arrange]{dplyr::arrange()}} before calling
\code{\link[dplyr:distinct]{dplyr::distinct()}} with the variables in \code{distinct_cols}. The variable is
arranged in descending order.}

\item{marginal_cols}{a subset of \code{group_cols} for which to marginally
summarize. If not supplied all grouping columns will be used. Use \code{NULL}
to not add any marginals.}

\item{add_reason_col}{whether or not to add a variable describing why an
observation was obfuscate, passed to \code{\link[=obfuscate_data]{obfuscate_data()}}}
}
\description{
Create aggregate data across all combinations, including totals.
Can specify if proportion or average should be used.
}
