% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_statistics.R
\name{get_data}
\alias{get_data}
\title{Get data from function call in `register`}
\usage{
get_data(register, fun, arguments = list())
}
\arguments{
\item{register}{character. Register short name.}

\item{fun}{character. function name from a register package}

\item{arguments}{list. additional arguments passed to `register::fun()`}
}
\value{
tibble
}
\description{
Data from either a installed `register` package for testing locally
or from the API when in active use
(controlled by global option `get_data_fun`)
}
