% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_vis.R
\name{get_measuring_frequency}
\alias{get_measuring_frequency}
\title{Gets the specified measure frequencies from a measure config}
\usage{
get_measuring_frequency(measure_config)
}
\arguments{
\item{measure_config}{the measure config from which to extract the measure
frequencies, the result of [search_measure_id()]}
}
\value{
vector of measure frequencies
}
\description{
Gets the specified measure frequencies from a measure config
}
