% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_vis.R
\name{get_organization_level}
\alias{get_organization_level}
\title{Gets the specified measure organizational levels from a measure config}
\usage{
get_organization_level(measure_config)
}
\arguments{
\item{measure_config}{the measure config from which to extract the measure
organizational levels, the result of [search_measure_id()]}
}
\value{
vector of measure organizational levels
}
\description{
Gets the specified measure organizational levels from a measure config
}
