% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_labels.R
\name{get_sos_vl}
\alias{get_sos_vl}
\title{Get value labels from socialstyrelsen}
\usage{
get_sos_vl(register = NULL)
}
\arguments{
\item{register}{Register at socialstyrelsen}
}
\description{
Get value labels from socialstyrelsen
}
