% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_surv.R
\name{get_surv_curve}
\alias{get_surv_curve}
\title{Gets the estimated survival curve, suitable for plotting}
\usage{
get_surv_curve(
  df,
  time_col = "time",
  event_col = "status",
  group_cols = NULL,
  marginal_cols,
  obfuscate_data = FALSE,
  add_reason_col = TRUE,
  censored_value = NA,
  censored_limit = 15,
  estimate = "survival"
)
}
\arguments{
\item{df}{data.frame}

\item{time_col}{the name of the time column}

\item{event_col}{the name of the event column}

\item{group_cols}{the columns to group by}

\item{marginal_cols}{the columns to add marginal groups for}

\item{obfuscate_data}{whether or not to obfuscate data}

\item{add_reason_col}{whether or not to add the reason for data being
censored}

\item{censored_value}{the value to use when observations are censored}

\item{censored_limit}{lower limir for at risk before censoring}

\item{estimate}{what the estimate should be, either "survival" or "event"}
}
\value{
estimated survival curve for each group
}
\description{
Like [get_surv_value()], but returns the full survival curve (one row
per event/censoring time, per group) instead of the estimate at a
single point in time.
}
