% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_surv.R
\name{get_surv_value}
\alias{get_surv_value}
\title{Gets the estimated survival at a specific point in time}
\usage{
get_surv_value(
  df,
  time,
  time_col = "time",
  event_col = "status",
  group_cols = NULL,
  marginal_cols
)
}
\arguments{
\item{df}{data.frame}

\item{time}{point in time (in days)}

\item{time_col}{the name of the time column}

\item{event_col}{the name of the event column}

\item{group_cols}{the columns to group by}

\item{marginal_cols}{the columns to add marginal groups for}
}
\value{
estimated survival at the requested point in time for each group
}
\description{
Gets the estimated survival at a specific point in time
}
