% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antal_andelar.R
\name{group_means}
\alias{group_means}
\title{Calculate n, means and sd by group}
\usage{
group_means(data, vars, ...)
}
\arguments{
\item{data}{A data.frame or tibble}

\item{vars}{Variables to calculate means and sd on}

\item{...}{Variables to group on by}
}
\description{
Calculate n, means and sd by group
}
