% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_statistics.R
\name{group_means}
\alias{group_means}
\title{Calculate n, means and sd by group}
\usage{
group_means(data, group_by, vars = NULL, obfuscate = TRUE, ...)
}
\arguments{
\item{data}{A data.frame or tibble}

\item{group_by}{Variables to group on by}

\item{vars}{Variables to calculate means and sd on.
Defaults to all vars in data.}

\item{obfuscate}{If data should be non-revealing}

\item{...}{Arguments passed to obfuscate}
}
\description{
Calculate n, means and sd by group
}
