% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_statistics.R
\name{group_proportions}
\alias{group_proportions}
\title{Proportions by group}
\usage{
group_proportions(data, ...)
}
\arguments{
\item{data}{A data.frame or tibble}

\item{...}{variables to group on}
}
\description{
Counts number of observations in each group.
drops last grouping variable and counts total.
Calculates proportion for each grouping
}
