% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locf.R
\name{locf}
\alias{locf}
\title{Performs memory efficient locf}
\usage{
locf(df, vars, groupby = "id", orderby = "date", slice = FALSE)
}
\arguments{
\item{df}{data.frame or tibble with data}

\item{vars}{variables to perform locf on}

\item{groupby}{variables to group on}

\item{orderby}{variable to sort each group on}

\item{slice}{keep only last row in each group}
}
\value{
data.frame or tibble with locf data
}
\description{
Performs memory efficient locf
}
