% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locf.R
\name{locfdt}
\alias{locfdt}
\title{locfdt}
\usage{
locfdt(
  dt,
  vars,
  groupby = "id",
  orderby = "date",
  slice = FALSE,
  return_tibble = FALSE,
  n_months = NULL
)
}
\arguments{
\item{dt}{data.frame or data.table to do LOCF imputation on}

\item{vars}{variables to impute with LOCF}

\item{groupby}{Variable to group on}

\item{orderby}{Variable to sort on}

\item{slice}{If `TRUE` will select the last row in each group of `groupby`}

\item{return_tibble}{whether to return a data.table or a tibble}

\item{n_months}{The maximum number of months the last observation can be
carried forward. If this is specified we recommend to also specify only the
variables of interest in `vars` to minimize the computational complexity.
If `n_months` is `NULL` LOCF will be carried out on all observations like
normal}
}
\value{
data.frame with LOCF imputation
}
\description{
DT Last Observation Carried Forward with
}
\details{
Returns data.frame with LOCF imputation using
data.table. Faster than `RCStat::locf` for large
datasets, especially when imputing on multiple variables
}
\examples{
\dontrun{df <- locfdt(df, vars)}

}
