% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rojande.R
\name{obfuscate_data}
\alias{obfuscate_data}
\title{Make data non-revealing}
\usage{
obfuscate_data(
  data,
  total_var = "total",
  count_var = "n",
  prop_var = "prop",
  group_var = NULL,
  statistics_vars = NULL,
  other_count_vars = NULL,
  round_statistics_vars = FALSE,
  round_statistics_digits = 2,
  add_reason_col = FALSE,
  liberal_obfuscation = TRUE,
  censored_value = 0,
  inform = TRUE,
  prop_scale = 1
)
}
\arguments{
\item{data}{The data.frame to obfuscate}

\item{total_var}{The column containing the denominator}

\item{count_var}{The column containing the numerator}

\item{prop_var}{The column containing the proportion as determined by
total_var and count_var}

\item{group_var}{The variables to group by before obfuscating. If data is
already grouped the grouping will be overwritten}

\item{statistics_vars}{Other statistics to be obfuscated such as mean,
standard deviation, etc. These are hidden if total_var is < 15}

\item{other_count_vars}{Other count vars to obfuscate. These have not been
used to calculate proportions. These are simply rounded to nearest 10}

\item{round_statistics_vars}{Whether or not to round statistics_vars}

\item{round_statistics_digits}{The number of digits statistics_vars are
rounded to. Passed to RCStat::roundc. Should be an integer or a list.
If it is a list it should a named list where each name is a column in
statistics_vars. The corresponding value in the list is the number of digits
to round to.}

\item{add_reason_col}{Whether or not to add a column called obfuscated_reason
indicating why a row was obfuscated with either "n < 5" or "N < 15" if
all of \code{count_var}, \code{total_var}, and \code{prop_var} are supplied. If only
\code{total_var} is found the only obfuscated_reason can be "N < 5"}

\item{liberal_obfuscation}{Whether or not to use liberal obfuscation. If
this is true we round proportions to 5\% when the numerator < 5.
We also do nothing when the numerator is < 5 and
the denominator is >= 200.}

\item{censored_value}{What to replace prop_var with when it is censored.
When working with bars it is useful to censor with 0 to not
affect order of bars but when producing a line plot 0 does not make sense.}

\item{inform}{Whether or not to print information about the current group
vars when \code{group_var} is \code{NULL}}

\item{prop_scale}{Argument indicating what scale the prop_var is. Either 1
or 100}
}
\description{
Make data non-revealing
}
\examples{

data_aggregated <- tibble::tribble(
  ~a, ~b, ~Count, ~Total,
  1,   1, 4,       26,
  1,   2, 16,      26,
  1,   1, 6,       26,
  2,   1, 24,      50,
  2,   2, 26,      50,
  3,   1, 7,       14,
  3,   2, 7,       14,
) |> dplyr::mutate(Proportion = Count/Total)

# Rows with less than 5 will be cleared
# frequencies rounded to nearest 10
# Proportion rounded to whole percents
# Groups with less than 15 in total are cleared
obfuscate_data(
  data_aggregated,
  count_var = "Count",
  total_var = "Total",
  prop_var = "Proportion"
)



data_aggregated <- tibble::tribble(
  ~a, ~b, ~n, ~x_non_missing, ~x_mean, ~x_sd,
  1,   1, 48, 12,             -0.593,   1.050,
  1,   2, 54, 16,             -0.308,   0.672,
  2,   1, 36, 31,              0.861,   0.502,
  2,   2, 66, 59,              0.401,   1.130,
)

# Clear statistics that is based on fewer than
# 15 observations
# Round statistics to 2 decimals
# Round frequencies, total and non_missing
obfuscate_data(
  data = data_aggregated,
  total_var = "x_non_missing",
  count_var = "n",
  statistics_vars = c("x_sd", "x_mean"),
  round_statistics_vars = TRUE
)

# Example with multiple variables in result
data_aggregated <- tibble::tribble(
  ~a, ~b, ~n, ~x_non_missing, ~y_non_missing, ~x_mean, ~x_sd, ~y_mean, ~y_sd,
  1,   1, 48, 42,             36,             -0.593,   1.050, 0.269,  0.214,
  1,   2, 54, 42,             30,             -0.308,   0.672, 0.176,  0.120,
  2,   1, 36, 30,             30,              0.861,   0.502, 0.465,  0.224,
  2,   2, 66, 60,             54,              0.401,   1.130, 0.557,  0.370
)
res <- data_aggregated

for (v in c("x", "y")) {
  res <- obfuscate_data(
    data = res,
    total_var = paste0(v, "_non_missing"),
    statistics_vars = c(paste0(v, "_sd"), paste0(v, "_mean")),
    round_statistics_vars = TRUE
  )
}

res


data_aggregated <- tibble::tribble(
  ~unit,  ~x_mean, ~x_sd,
  1,      1.0503,   0.26956,
  1,      0.6725,   0.17612,
  2,      0.5021,   0.46552,
  2,      1.1307,   0.55783
)

res <- obfuscate_data(
  data = data_aggregated,
  statistics_vars = c("x_mean", "x_sd"),
  round_statistics_digits = list(x_mean = 2, x_sd = 1),
  round_statistics_vars = TRUE
)

res
}
