% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_statistics.R
\name{pivot_prop_count}
\alias{pivot_prop_count}
\title{Pivots the result of a single prop_count from [get_aggregate_value()]}
\usage{
pivot_prop_count(df, category_name = "kategori")
}
\arguments{
\item{df}{the data.frame to plot}

\item{category_name}{the name of the variable to store the outcomes of the
variable in}
}
\value{
a data.frame pivoted into long format
}
\description{
Pivots the result of a single prop_count from [get_aggregate_value()]
}
