% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_vis.R
\name{postprocess_indicator}
\alias{postprocess_indicator}
\alias{postprocess_indicator_prop}
\alias{postprocess_indicator_median}
\alias{postprocess_indicator_count}
\alias{postprocess_indicator_org}
\alias{postprocess_indicator_register}
\alias{postprocess_indicator_names}
\alias{postprocess_indicator_reason}
\title{Handles the necessary postprocessing of an aggregated indicator by renaming
and adding necessary variables as well as adding the version}
\usage{
postprocess_indicator(df, cfg, register_id)

postprocess_indicator_prop(df)

postprocess_indicator_median(df)

postprocess_indicator_count(df)

postprocess_indicator_org(df, register_id)

postprocess_indicator_register(df, register_id)

postprocess_indicator_names(df)

postprocess_indicator_reason(df)
}
\arguments{
\item{df}{an aggregated indicator}

\item{cfg}{the measure config, the result of [get_measure_config()]}

\item{register_id}{the register ID}
}
\value{
renamed data.frame
}
\description{
Handles the necessary postprocessing of an aggregated indicator by renaming
and adding necessary variables as well as adding the version
}
\section{Functions}{
\itemize{
\item \code{postprocess_indicator_prop()}: renames columns if the indicator is a
proportion

\item \code{postprocess_indicator_median()}: renames columns if the indicator is a
median

\item \code{postprocess_indicator_count()}: renames columns if the indicator is a
count

\item \code{postprocess_indicator_org()}: renames county and unit variables

\item \code{postprocess_indicator_register()}: adds the register name and HSAID variables

\item \code{postprocess_indicator_names()}: adds remaining variables

\item \code{postprocess_indicator_reason()}: adds ReasonCode

}}
