% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prettify_table}
\alias{prettify_table}
\title{Prettifies a table}
\usage{
prettify_table(df, vars = NULL, remove_vars = NULL, ...)
}
\arguments{
\item{df}{data.frame to prettify}

\item{vars}{list of variable names, passed to [RCStat::decode_vars()]}

\item{remove_vars}{vector of variable names to remove from the table}

\item{...}{manual renames, passed to [dplyr::rename()] and fully supports
tidy-selection of columns}
}
\value{
prettified data.frame
}
\description{
Renames columns, rounds means, medians, and proportions, and multiplies
proportions. Also handles output from [get_surv_value()] and
[get_surv_curve()] (`time`, `estimate`, `n.risk`, `cum_events`)
}
