% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prettify_table}
\alias{prettify_table}
\title{Prettifies a table}
\usage{
prettify_table(df, vars = NULL, ...)
}
\arguments{
\item{df}{data.frame to prettify}

\item{vars}{list of variable names, passed to [RCStat::decode_vars()]}

\item{...}{manual renames, passed to [dplyr::rename()] and fully supports
tidy-selection of columns}
}
\value{
prettified data.frame
}
\description{
Renames columns, rounds means, medians, and proportions, and multiplies
proportions
}
