% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_password}
\alias{random_password}
\title{Random password}
\usage{
random_password(char_types = c("R", "N", "!"), password_length = 15)
}
\arguments{
\item{char_types}{Possible character types to use in password as vector.
Documentation on API:
https://www.passwordrandom.com/pronounceable-password-generator
     Syntax:
       L - upper case letter; l (small L) - lower case letter;
       R or r - random case letter;
       N or n - number (0-9);
       # - symbol (!@#$.+);
       ! - all symbols;
       C - upper case consonant letter;
       c - lower case consonant letter;
       V - upper case vowel letter;
       v - lower case vowel letter;}

\item{password_length}{integer, number of characters in password}
}
\description{
Gets random password from API
}
\examples{
\dontrun{rand_pass <- random_password(c("R", "N"), 10)}
}
