% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCStat-package.R
\docType{package}
\name{RCStat-package}
\alias{RCStat}
\alias{RCStat-package}
\title{RCStat: Commonly used}
\description{
Holds functions for standard tables and general data management for registers at Registercentrum.
}
\section{Package options}{

\itemize{
\item \code{rcstat.local_plumber}: logical, default \code{FALSE}.
Controls whether \code{\link[=api_statistics]{api_statistics()}} should send API requests to a local
plumber instance.
Can be set globally with \code{options(rcstat.local_plumber = TRUE)}.
\item \code{rcstat.local_scope}: integer, default 1.
Controls the scope of the API call.
\item \code{rcstat.local_unit}: integer, default 0.
Controls what unit you are authenticated for.
\item \code{rcstat.local_role}: integer, default 1.
Controls the role you are authenticates as.
\item \code{rcstat.local_register}: integer, default 1.
Controls the register you are authenticated for.
}
}

\keyword{internal}
