% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCStat-package.R
\docType{package}
\name{RCStat-package}
\alias{RCStat}
\alias{RCStat-package}
\title{RCStat: Commonly used}
\description{
Holds functions for standard tables and general data management for registers at Registercentrum.
}
\keyword{internal}
