% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCStat-package.R
\docType{package}
\name{RCStat-package}
\alias{RCStat}
\alias{RCStat-package}
\title{RCStat: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
