% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_vis.R
\name{search_submeasures}
\alias{search_submeasures}
\title{Search for submeasures related to a measure}
\usage{
search_submeasures(id)
}
\arguments{
\item{id}{the key `id` as returned by [search_measure_id], separate from a
MeasureID}
}
\value{
the result of the API call
}
\description{
Search for submeasures related to a measure
}
