% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sos_metadata}
\alias{sos_metadata}
\title{Saves data.frames as .csv files and creates and saves the corresponding
metadata files. Can also zip the outputs with a random password which is
saved in a separate .txt file.}
\usage{
sos_metadata(
  dfs = list(),
  file_names = list(),
  separator = ",",
  zip = TRUE,
  output_dir = "Output",
  zip_file_name = "output"
)
}
\arguments{
\item{dfs}{A list of data.frames to export}

\item{file_names}{A list of the names to use when saving each data.frame in
`dfs` without file extension}

\item{separator}{The separator to use when saving the `dts`}

\item{zip}{Whether or not to zip the output directory after saving the files}

\item{output_dir}{The output directory in which to save the files}

\item{zip_file_name}{The name of the created zip file without file extension}
}
\description{
Saves data.frames as .csv files and creates and saves the corresponding
metadata files. Can also zip the outputs with a random password which is
saved in a separate .txt file.
}
\examples{
\dontrun{
data_1 <- RCDBT::GetRegisterData(...) \%>\% filter(...)
data_2 <- RCDBT::GetRegisterData(...) \%>\% filter(...)
sos_metadata(dfs = list(data_1, data_2),
             file_names = list("dnr_xxxx_yyyyy_1", "dnr_xxxx_yyyyy_2"))
}
}
