% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_api.R
\name{statistics_api}
\alias{statistics_api}
\title{Get data from statistics API}
\usage{
statistics_api(register, script, arguments = NULL)
}
\arguments{
\item{register}{register short name}

\item{script}{script name}

\item{arguments}{Arguments to include as named list}
}
\description{
Get data from statistics API
}
