% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_labels.R
\name{var_to_char}
\alias{var_to_char}
\title{Convert variable to character}
\usage{
var_to_char(x, labels, name = NULL)
}
\arguments{
\item{x}{Object with "map" attributes, where levels and labels exist}

\item{name}{name of input variable (optional)}

\item{droplevels}{If `TRUE`, factor levels not present in data will be
excluded.}
}
\value{
Factor
}
\description{
Used to extract texts from a variable containing numeric values.
}
\examples{
df <- data.frame(cars = 1:3)
labels <- data.frame(
    levels = 1:3,
    labels = c("Volvo", "Saab", "Opel")
 )
var_to_char(df[["cars"]])
}
