% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_labels.R
\name{var_to_factor}
\alias{var_to_factor}
\title{Convert object with attributes in "map" to factor}
\usage{
var_to_factor(x, labels, name = NULL, droplevels = TRUE)
}
\arguments{
\item{x}{Vector with codes to decode}

\item{labels}{data.frame with columns ValueCode and ValueName to use
when decoding.}

\item{name}{name of input variable (optional). Used in the warning message
only to print the name of variable with missing labels.}

\item{droplevels}{If `TRUE`, factor levels not present in data will be
excluded.}
}
\value{
Factor
}
\description{
Used to extract texts from a variable containing numeric values.
}
\examples{

x <- sample(1:3, size = 10, replace = TRUE)

labels <- data.frame(
  ValueCode = 1:3,
  ValueName = c("Volvo", "Saab", "Opel")
)
var_to_factor(x, labels, name = "cars")
}
