% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_labels.R
\name{var_to_factor}
\alias{var_to_factor}
\title{Convert object with attributes in "map" to factor}
\usage{
var_to_factor(x, labels)
}
\arguments{
\item{x}{Object with "map" attributes, where levels and labels exist}

\item{droplevels}{If `TRUE`, factor levels not present in data will be
excluded.}

\item{name}{name of input variable (optional)}
}
\value{
Factor
}
\description{
Used to extract texts from a variable containing numeric values.
}
\examples{
df <- data.frame(cars = 1:3)
attr(df$cars, "map") <- data.frame(
    levels = 1:3,
    labels = c("Volvo", "Saab", "Opel")
 )
attr_to_factor(df$cars)
}
