% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zip_dir_with_pass}
\alias{zip_dir_with_pass}
\title{Zip directory with password}
\usage{
zip_dir_with_pass(
  directory = "output",
  pass = random_password(),
  file = "output",
  overwrite = TRUE
)
}
\arguments{
\item{directory}{directory to zip. Defaults to `output`}

\item{pass}{password to use. Defaults to `random_password()`}

\item{file}{name of zip-file}

\item{overwrite}{Overwrite zip-file if it exists}
}
\description{
zips chosen directory with password protection. Zip-file and
txt-file with the password are placed in working directory
}
\examples{
\dontrun{
zip_dir_with_pass()
}
}
